♻️ split artifact preprocess source broker responsibilities
