+++ apps/opencrane/README.md
@@ -77,6 +77,8 @@ its resources to the lifecycle owner.
 - `src/app/routes.ts` contains only named per-area route lists and the trivial mount loop.
 - `src/app/runtime-composition.ts` binds workload identity review, durable dispatch, and external-I/O
   ports without choosing transport paths.
+- `src/infra/artifacts/artifact-preprocess-source-broker.factory.ts` composes the fenced source-byte
+  broker from separately owned claim, mounted-key, and private read-port responsibilities.
 - `src/app/background-workers.ts` owns schedule ticks, expired-run repair, and fenced cleanup loops.
 - `src/app/lifecycle.ts` starts both listeners, stops producers first, drains requests, disconnects
   Prisma, and flushes telemetry.
