+++ apps/opencrane/src/infra/artifacts/artifact-preprocess-source-broker.factory.ts
@@ -1,0 +1,62 @@
+import type { PrismaClient } from "@prisma/client";
+
+import { PrismaArtifactPreprocessRepository, type ArtifactPreprocessSourceBroker } from "@opencrane/backend/server/agents/artifacts";
+import { ___DoWithTrace } from "@opencrane/observability";
+
+import { _CreateArtifactReadLeaseSigner } from "./artifact-read-lease-signer.factory.js";
+import { _CreateArtifactServiceReadPort, _InternalArtifactServiceUrl } from "./artifact-service-read-port.factory.js";
+
+/**
+ * Compose the server-side preprocessing source broker from the named
+ * catalogue, signing, and byte-transport responsibilities.
+ *
+ * The returned broker exposes source bytes only after the durable claim fence
+ * creates exact, expiring read claims. Neither the worker nor the router sees
+ * the storage coordinate or signed read lease.
+ * @param prisma - Canonical product-authority database client.
+ * @param environment - Server configuration containing service and key paths.
+ * @returns Source broker for the workload-authenticated preprocessor router.
+ */
+export function _CreateArtifactPreprocessSourceBroker(prisma: PrismaClient, environment: NodeJS.ProcessEnv = process.env): ArtifactPreprocessSourceBroker
+{
+	const jobs = new PrismaArtifactPreprocessRepository(prisma);
+	const serviceUrl = _InternalArtifactServiceUrl(environment.ARTIFACT_SERVICE_URL ?? "");
+	const signLease = _CreateArtifactReadLeaseSigner(environment);
+	const readPort = _CreateArtifactServiceReadPort(serviceUrl);
+	return _CreateArtifactPreprocessSourceReader(jobs, signLease, readPort);
+}
+
+/**
+ * Create the operation that turns one current preprocess claim into verified
+ * server-brokered source bytes.
+ *
+ * It owns the post-transaction expiry check and metadata cross-check. It does
+ * not choose a revision, extend a claim, or expose a lease outside this server.
+ * @param jobs - Durable source-lease issuer guarded by the current claim fence.
+ * @param signLease - Server-only signer for the exact issued read claims.
+ * @param readPort - Private artifact-service byte transport.
+ * @returns Fence-aware source byte reader.
+ */
+function _CreateArtifactPreprocessSourceReader(jobs: PrismaArtifactPreprocessRepository, signLease: ReturnType<typeof _CreateArtifactReadLeaseSigner>, readPort: ReturnType<typeof _CreateArtifactServiceReadPort>): ArtifactPreprocessSourceBroker
+{
+	return {
+		async read(command)
+		{
+			return ___DoWithTrace("artifact-preprocessor.source.broker", { jobId: command.jobId, attempt: command.attempt }, async function _ReadSource()
+			{
+				// 1. Allocate exact read claims under the current database-owned fence and its old deadline.
+				const source = await jobs.issueSourceLeaseAtomically(command);
+				if (source === null) return null;
+
+				// 2. Refuse a claim that expired after the transaction, then sign without extending its authority.
+				if (source.readLease.expiresAtEpochSeconds <= Math.floor(Date.now() / 1_000)) return null;
+				const compactLease = signLease(source.readLease);
+
+				// 3. Cross-check storage metadata before proxying bytes without exposing the lease.
+				const response = await readPort.read(compactLease);
+				if (response.body === null || response.headers.get("content-length") !== String(source.byteLength) || response.headers.get("content-type") !== source.mediaType) throw new Error("artifact service read metadata did not match the claimed source");
+				return { byteLength: source.byteLength, mediaType: source.mediaType, bytes: response.body as unknown as AsyncIterable<Uint8Array> };
+			});
+		},
+	};
+}
