+++ libs/backend/server/agents/artifacts/main/src/artifact-read-lease.ts
@@ -2,7 +2,7 @@ import { randomUUID } from "node:crypto";
 
 import { __IsSafeArtifactMediaType } from "@opencrane/backend/artifacts/authorization";
 
-import type { ArtifactReadLeaseRepository, ArtifactReadLeaseSigner, IssueArtifactReadLeaseCommand, IssueArtifactReadLeaseResult, PublishedArtifactReadTarget } from "./artifact-read-lease.types.js";
+import { IssueArtifactReadLeaseOutcomes, type ArtifactReadLeaseRepository, type ArtifactReadLeaseSigner, type IssueArtifactReadLeaseCommand, type IssueArtifactReadLeaseResult, type PublishedArtifactReadTarget } from "./artifact-read-lease.types.js";
 
 /** Maximum lifetime for a service-verified internal artifact read lease. */
 const _READ_LEASE_SECONDS = 300;
@@ -12,21 +12,21 @@ export async function __IssueArtifactReadLease(repository: ArtifactReadLeaseRepo
 {
 	if (!_IsCommand(command) || !Number.isSafeInteger(nowEpochSeconds) || nowEpochSeconds < 0)
 	{
-		return { outcome: "denied", reason: "invalid_command" };
+		return { outcome: IssueArtifactReadLeaseOutcomes.Denied, reason: "invalid_command" };
 	}
 
 	// 1. Reload the exact revision under its silo; no caller byte metadata becomes lease content.
 	const target = await repository.loadPublishedReadTarget(command);
 	if (!_MatchesCommand(target, command))
 	{
-		return { outcome: "denied", reason: "revision_not_readable" };
+		return { outcome: IssueArtifactReadLeaseOutcomes.Denied, reason: "revision_not_readable" };
 	}
 
 	// 2. Bind one opaque, bounded lease to the immutable facts returned by the catalogue authority.
 	const claims = { leaseId: randomUUID(), siloId: target.siloId, artifactId: target.artifactId, artifactRevisionId: target.artifactRevisionId, contentAddress: target.contentAddress, byteLength: target.byteLength, mediaType: target.mediaType, action: "artifact.read" as const, expiresAtEpochSeconds: nowEpochSeconds + _READ_LEASE_SECONDS };
 
 	// 3. Only app composition owns the mounted signing key and may serialize the reviewed claims.
-	return { outcome: "issued", compactLease: signer.sign(claims), claims };
+	return { outcome: IssueArtifactReadLeaseOutcomes.Issued, compactLease: signer.sign(claims), claims };
 }
 
 /** Reject malformed durable coordinates before they can probe the catalogue repository. */
