+++ libs/backend/server/agents/artifacts/main/src/index.ts
@@ -2,6 +2,7 @@ export { __FinalizeArtifactRevision } from "./artifact-finalization.js";
 export { __ClaimArtifactPreprocessJob, __CompleteArtifactPreprocessJob, __FailArtifactPreprocessJob, __IssueArtifactPreprocessOutputLease } from "./artifact-preprocessing.js";
 export { __CreateArtifactPreprocessorRouter } from "./artifact-preprocessing.router.js";
 export { __IssueArtifactReadLease } from "./artifact-read-lease.js";
+export { IssueArtifactReadLeaseOutcomes } from "./artifact-read-lease.types.js";
 export { PrismaArtifactAuthorityRepository } from "./prisma-artifact-authority.js";
 export { PrismaArtifactPreprocessRepository } from "./prisma-artifact-preprocessing.js";
 export { __UploadArtifact } from "./artifact-upload.js";
